+++ src/LeapInstrument.java
@@ -2,19 +2,19 @@ public interface LeapInstrument {
 	
 	public void playNote(String note, int octave) throws Exception;
 	
-	public void playNote(String note, int octave, int vol) throws Exception;
+	public void playNote(String note, int octave, double vol) throws Exception;
 	
 	public void playNote(int note, int octave) throws Exception;
 	
-	public void playNote(int note, int octave, int vol) throws Exception;
+	public void playNote(int note, int octave, double vol) throws Exception;
 	
 	public void playDrum(String drum) throws Exception;
 	
-	public void playDrum(String drum, int vol) throws Exception;
+	public void playDrum(String drum, double vol) throws Exception;
 	
-	public int getVol();
+	public double getVol();
 	
-	public int setVol(int vol);
+	public double setVol(double vol);
 	
 	public int getChannel();
